% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censored_cumulative.R
\name{censored_cumulative_logit}
\alias{censored_cumulative_logit}
\alias{censored_cumulative_logit_stanvars}
\title{Censored cumulative logistic family.}
\usage{
censored_cumulative_logit()

censored_cumulative_logit_stanvars()
}
\description{
Use as:
brm(min | vint(max)  ~ predictors , family = censored_cumulative_logit(), stanvars = censored_cumulative_logit_stanvars(), ...)
Non-censored observations mean min == max.
}
